chore: dir structure fix
