+++ src/components/common/common.css
@@ -1,0 +1,20 @@
+* {
+    box-sizing: border-box;
+    font-family: Pretendard;
+}
+
+:root {
+    --primary-color: #6D6AFE;
+    --red-color: #FF5B56;
+    --black-color: #111322;
+    --white-color: #ffffff;
+    --gray100-color: #3e3e43;
+    --gray80-color: #9fa6b2;
+    --gray60-color: #ccd5e3;
+    --gray40-color: #e7effb;
+    --gray20-color: #f0f6ff;
+}
+
+a {
+    text-decoration: none;
+}
