+++ src/components/common/header.css
@@ -1,0 +1,39 @@
+.nav {
+    width: 100%;
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    background-color: #F0F6FF;
+    padding: 33px 200px 32px;
+}
+
+.link {
+    width: 128px;
+    background: linear-gradient(91deg, #6D6AFE 0.12%, #6AE3FE 101.84%);
+    padding: 16px 20px;
+    border-radius: 8px;
+    color: #F5F5F5;
+    font-size: 18px;
+    font-weight: 600;
+    text-align: center;
+}
+
+.account {
+    display: flex;
+    gap: 6px;
+    justify-content: center;
+    align-items: center;
+}
+
+.account-img {
+    width: 28px;
+    height: 28px;
+    border-radius: 20px;
+}
+
+/* tablet */
+@media (max-width:1124px) {
+    .nav {
+        padding: 33px 32px 32px;
+    }
+}
