--- src/css/footer.css
@@ -1,23 +1,0 @@
-.footer-box {
-    height: 160px;
-    padding: 32px 104px 108px;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    background: #111322;
-    color: #676767;
-}
-
-.faq{
-    display: flex;
-    gap: 30px;
-}
-
-.faq a{
-    color: #676767; 
-}
-
-.sns {
-    display: flex;
-    gap: 12px ;
-}
