--- src/css/main.css
@@ -1,99 +1,0 @@
-.profile {
-    background-color: #F0F6FF;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    padding: 20px 60px;
-    gap: 20px;
-    text-align: center;
-}
-
-.profile-img {
-    width: 60px;
-    height: 60px;
-    border-radius: 47px;
-    margin: 0 auto;
-}
-
-.profile-img-name {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    gap: 12px;
-    word-break: keep-all;
-}
-
-.profile-name {
-    font-size: 40px;
-    font-weight: 600;
-    word-break: keep-all;
-}
-
-.main {
-    padding: 40px 0;
-    background-color: #fff;
-    display: flex;
-}
-
-.main-box {
-    margin: 0 auto;
-    display: flex;
-    gap: 40px;
-    flex-direction: column;
-}
-
-.search-box {
-    padding: 15px 16px;
-    border-radius: 10px;
-    background: #F5F5F5;
-    display: flex;
-    align-items: center;
-    gap:10px;
-}
-
-.search-input {
-    width: 1000px;
-    border: none;
-    background: #F5F5F5;
-}
-
-input::placeholder{
-    color: #666;
-}
-
-.cards-grid {
-    display: grid;
-    grid-template-columns: repeat(3, 1fr);
-    gap: 25px 20px;
-}
-
-/* mobile */
-@media  (max-width:767px) {
-    .cards-grid {
-        grid-template-columns: 1fr;
-        gap: 20px 0;
-        margin: 0 32px;
-        align-items: center;
-        justify-items: center;
-    }
-
-    .search-input {
-        width: 375px;
-    }
-}
-
-/* tablet */
-@media (min-width:768px) and (max-width:1124px) {
-    .cards-grid {
-        grid-template-columns: repeat(2, 1fr);
-        gap: 24px 25px;
-        margin: 0 32px;
-        align-items: center;
-        justify-items: center;
-    }
-
-    .search-input {
-        width: 700px;
-    }
-}
-
